+++ adapter/src/index.js
@@ -15,6 +15,10 @@ const AppAdapter = ({
     apiVersion,
     pwaEnabled,
     plugin,
+    parentAlertsAdd,
+    showAlertsInPlugin,
+    onPluginError,
+    clearPluginError,
     loginApp,
     children,
 }) => {
@@ -25,6 +29,7 @@ const AppAdapter = ({
                 onRetry={() => {
                     window.location.reload()
                 }}
+                plugin={false}
             >
                 <ServerVersionProvider
                     appName={appName}
@@ -32,7 +37,7 @@ const AppAdapter = ({
                     url={url}
                     apiVersion={apiVersion}
                     pwaEnabled={pwaEnabled}
-                    loginApp={true}
+                    plugin={false}
                 >
                     <LoginAppWrapper>{children}</LoginAppWrapper>
                 </ServerVersionProvider>
@@ -40,7 +45,11 @@ const AppAdapter = ({
         )
     }
     return (
-        <ErrorBoundary fullscreen onRetry={checkForSWUpdateAndReload}>
+        <ErrorBoundary
+            plugin={plugin}
+            fullscreen
+            onRetry={checkForSWUpdateAndReload}
+        >
             <OfflineInterfaceProvider>
                 <PWALoadingBoundary>
                     <ServerVersionProvider
@@ -49,9 +58,15 @@ const AppAdapter = ({
                         url={url}
                         apiVersion={apiVersion}
                         pwaEnabled={pwaEnabled}
-                        loginApp={false}
+                        plugin={plugin}
+                        parentAlertsAdd={parentAlertsAdd}
+                        showAlertsInPlugin={showAlertsInPlugin}
                     >
-                        <AppWrapper plugin={plugin} loginApp={loginApp}>
+                        <AppWrapper
+                            plugin={plugin}
+                            onPluginError={onPluginError}
+                            clearPluginError={clearPluginError}
+                        >
                             {children}
                         </AppWrapper>
                     </ServerVersionProvider>
@@ -66,10 +81,14 @@ AppAdapter.propTypes = {
     appVersion: PropTypes.string.isRequired,
     apiVersion: PropTypes.number,
     children: PropTypes.element,
+    clearPluginError: PropTypes.func,
     loginApp: PropTypes.bool,
+    parentAlertsAdd: PropTypes.func,
     plugin: PropTypes.bool,
     pwaEnabled: PropTypes.bool,
+    showAlertsInPlugin: PropTypes.func,
     url: PropTypes.string,
+    onPluginError: PropTypes.func,
 }
 
 export default AppAdapter