+++ pwa/src/service-worker/utils.js
@@ -11,6 +11,11 @@ const APP_ADAPTER_URL_PATTERNS = [
     /\/api(\/\d+)?\/userSettings/, // useLocale
     /\/api(\/\d+)?\/me\?fields=id$/, // useVerifyLatestUser
 ]
+// Note that the CRA precache manifest files start with './'
+// TODO: Make this extensible with a d2.config.js option
+export const CRA_MANIFEST_EXCLUDE_PATTERNS = [
+    /^\.\/static\/js\/moment-locales\//,
+]
 
 // '[]' Fallback prevents error when switching from pwa enabled to disabled
 const APP_SHELL_URL_FILTER_PATTERNS = JSON.parse(
@@ -52,6 +57,14 @@ export function setUpKillSwitchServiceWorker() {
 }
 
 export function urlMeetsAppShellCachingCriteria(url) {
+    // If this request is for a file that belongs to this app, cache it
+    // (in production, many, but not all, app files will be precached -
+    // e.g. moment-locales is omitted)
+    const appScope = new URL('./', self.location.href)
+    if (url.href.startsWith(appScope.href)) {
+        return true
+    }
+
     // Cache this request if it is important for the app adapter to load
     const isAdapterRequest = APP_ADAPTER_URL_PATTERNS.some((pattern) =>
         pattern.test(url.href)
@@ -107,18 +120,11 @@ export async function removeUnusedCaches() {
     )
 }
 
-/**
- * Can be used to access information about this service worker's clients.
- * Sends back information on a message with 'CLIENTS_INFO' type; the payload
- * currently contains the number of current clients, including uncontrolled.
- * @returns {Object} { clientsCounts: number }
- */
-export async function getClientsInfo(event) {
-    const clientId = event.source.id
-
+/** Get all clients including uncontrolled, but only those within SW scope */
+export function getAllClientsInScope() {
     // Include uncontrolled clients: necessary to know if there are multiple
     // tabs open upon first SW installation
-    const filteredClientsList = await self.clients
+    return self.clients
         .matchAll({
             includeUncontrolled: true,
         })
@@ -129,12 +135,24 @@ export async function getClientsInfo(event) {
                 client.url.startsWith(self.registration.scope)
             )
         )
+}
+
+/**
+ * Can be used to access information about this service worker's clients.
+ * Sends back information on a message with 'CLIENTS_INFO' type; the payload
+ * currently contains the number of current clients, including uncontrolled.
+ * @returns {Object} { clientsCounts: number }
+ */
+export async function getClientsInfo(event) {
+    const clientId = event.source.id
+
+    const clientsList = await getAllClientsInScope()
 
     self.clients.get(clientId).then((client) => {
         client.postMessage({
             type: swMsgs.clientsInfo,
             payload: {
-                clientsCount: filteredClientsList.length,
+                clientsCount: clientsList.length,
             },
         })
     })